List, create and edit plans

+++ packages/api/src/migrations/20210913110038_create_plans_table.js
@@ -3,7 +3,7 @@ export async function up(knex) {
     return knex.schema.createTable('plans', table => {
         table.increments('id').primary();
         table.string('name');
-        table.decimal('amount', 3, 2);
+        table.float('amount', 3, 2);
         table.timestamps();
     });
 }
